+++ utils/const.py
@@ -2,7 +2,7 @@
 
 from enum import StrEnum, unique
 
-Version = "1.5.0"  # pylint: disable=invalid-name  # Mayor.minor.patch; usado por App_version()
+Version = "1.6.0"  # pylint: disable=invalid-name  # Mayor.minor.patch; usado por App_version()
 
 # Tiempos por defecto (segundos)
 DEFAULT_AWAY_TIME = 2